# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sensor Management API Documentation",
        "description": "API Documentation for Sensor Management System",
        "contact": {
            "name": "",
            "email": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/sensors/{sensorId}": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Get sensor details",
                "description": "Returns details for a specific sensor",
                "operationId": "getSensor",
                "parameters": [
                    {
                        "name": "sensorId",
                        "in": "path",
                        "description": "ID of the sensor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SensorResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sensor not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Update an existing sensor",
                "description": "Updates a sensor in the system",
                "operationId": "updateSensor",
                "parameters": [
                    {
                        "name": "sensorId",
                        "in": "path",
                        "description": "ID of the sensor to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Sensor data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SensorUpdateRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SensorUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "404": {
                        "description": "Sensor not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Delete a sensor",
                "description": "Deletes a sensor from the system",
                "operationId": "deleteSensor",
                "parameters": [
                    {
                        "name": "sensorId",
                        "in": "path",
                        "description": "ID of the sensor to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "404": {
                        "description": "Sensor not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sensors": {
            "post": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Create a new sensor",
                "description": "Creates a new sensor in the system",
                "operationId": "createSensor",
                "requestBody": {
                    "description": "Sensor data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SensorStoreRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SensorStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SensorStoreRequest": {
                "required": [
                    "sensor_name",
                    "type",
                    "display_name",
                    "is_output_binary",
                    "color_class",
                    "sensor_group_id"
                ],
                "properties": {
                    "sensor_name": {
                        "description": "Name of the sensor",
                        "type": "string",
                        "maxLength": 50
                    },
                    "type": {
                        "description": "Type identifier of the sensor",
                        "type": "string",
                        "maxLength": 50
                    },
                    "display_name": {
                        "description": "Display name of the sensor",
                        "type": "string",
                        "maxLength": 50
                    },
                    "unit_of_measurement": {
                        "description": "Unit of measurement for sensor readings",
                        "type": "string",
                        "maxLength": 10,
                        "nullable": true
                    },
                    "is_output_binary": {
                        "description": "Whether sensor output is binary",
                        "type": "boolean"
                    },
                    "color_class": {
                        "description": "CSS color class for the sensor",
                        "type": "string"
                    },
                    "image": {
                        "description": "Sensor image file",
                        "type": "string",
                        "format": "binary"
                    },
                    "icon": {
                        "description": "Sensor icon file",
                        "type": "string",
                        "format": "binary"
                    },
                    "sensor_group_id": {
                        "description": "ID of the sensor group",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "SensorUpdateRequest": {
                "required": [
                    "sensor_name",
                    "type",
                    "display_name",
                    "is_output_binary",
                    "color_class",
                    "sensor_group_id"
                ],
                "properties": {
                    "sensor_name": {
                        "description": "Name of the sensor",
                        "type": "string",
                        "maxLength": 50
                    },
                    "type": {
                        "description": "Type identifier of the sensor",
                        "type": "string",
                        "maxLength": 50
                    },
                    "display_name": {
                        "description": "Display name of the sensor",
                        "type": "string",
                        "maxLength": 50
                    },
                    "unit_of_measurement": {
                        "description": "Unit of measurement for sensor readings",
                        "type": "string",
                        "maxLength": 10,
                        "nullable": true
                    },
                    "is_output_binary": {
                        "description": "Whether sensor output is binary",
                        "type": "boolean"
                    },
                    "color_class": {
                        "description": "CSS color class for the sensor",
                        "type": "string"
                    },
                    "image": {
                        "description": "Sensor image file",
                        "type": "string",
                        "format": "binary"
                    },
                    "icon": {
                        "description": "Sensor icon file",
                        "type": "string",
                        "format": "binary"
                    },
                    "sensor_group_id": {
                        "description": "ID of the sensor group",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "EnumResource": {
                "properties": {
                    "value": {
                        "description": "Enum value",
                        "type": "string"
                    },
                    "label": {
                        "description": "Human-readable label",
                        "type": "string"
                    },
                    "symbol": {
                        "description": "Symbol representation",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "SensorResource": {
                "properties": {
                    "id": {
                        "description": "Sensor ID",
                        "type": "integer"
                    },
                    "sensor_name": {
                        "description": "Name of the sensor",
                        "type": "string"
                    },
                    "type": {
                        "description": "Type identifier of the sensor",
                        "type": "string"
                    },
                    "display_name": {
                        "description": "Display name of the sensor",
                        "type": "string"
                    },
                    "unit_of_measurement": {
                        "description": "Unit of measurement for sensor readings",
                        "type": "string",
                        "nullable": true
                    },
                    "is_output_binary": {
                        "description": "Whether sensor output is binary",
                        "type": "boolean"
                    },
                    "image_path": {
                        "description": "Path to sensor image",
                        "type": "string",
                        "nullable": true
                    },
                    "icon_path": {
                        "description": "Path to sensor icon",
                        "type": "string",
                        "nullable": true
                    },
                    "color_class": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/EnumResource"
                            }
                        ],
                        "nullable": true
                    },
                    "group_name": {
                        "description": "Name of the sensor group",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sensorAdminKey": {
                "type": "apiKey",
                "name": "X-Admin-Token",
                "in": "header"
            }
        }
    },
...
}